@@ -1,4 +1,4 @@
-const  { mongoose, Schema, ObjectId } = require("../Discord Bot/node_modules/mongoose");
+const { mongoose, Schema, ObjectId } = require("../Discord Bot/node_modules/mongoose");
 
 const Setting = new Schema({
 	id: ObjectId,
@@ -8,7 +8,6 @@ const Setting = new Schema({
 	pollRole: { type: String, default: null },
 	deleteMoviesRole: { type: String, default: null },
 	viewedMoviesRole: { type: String, default: null },
-	//If deleteMoviesRole = null, allow only admins delete. = "all" then remove restrictions. If specific role then admins + role
 });
 
 module.exports = mongoose.model("Settings", Setting);
