@@ -1,5 +1,6 @@
-export { ErrorComponent } from './error';
-export { FooterComponent } from './footer';
-export { Loader } from './loader';
-export { MovieComponent } from './movie';
-export { NavbarComponent } from './navbar';
+export { ErrorComponent } from "./error";
+export { FooterComponent } from "./footer";
+export { Loader } from "./loader";
+export { MovieComponent } from "./movie";
+export { NavbarComponent } from "./navbar";
+export { Command } from "./command";
